Update README.md
In Eng
This research aims to study and analyze the conditions of a hailstorm that occurred in the Chiang Khong district of Chiang Rai Province on April 23, 2020. Dual-polarimetric radar data from the Thai Meteorological Department was utilized to examine the structure of the rain clouds that developed during the event. The movement of the air masses was analyzed using ERA5 reanalysis data. Simulated atmospheric data from ERA5 in the vicinity of the affected area were created to classify hailstones during the incident. Subsequently, the geographical features related to wind and the movement of rain cloud groups were analyzed. The classification of hailstones was performed using the CSU Tool method, and PyTDA and Single Dop were used to analyze wind patterns. If there are any questions, please contact Assistant Professor Dr. Nattapon Mahavik at nattaponm@nu.ac.th

In Thai
งานวิจัยชิ้นนี้ มุ่งศึกษาและวิเคราะห์สภาวะลูกเห็บ ที่เกิดในพื้นที่จังหวัดเชียงราย อำเภอเชียงของ ในวันที่ 23 เมษายน 2020 ใช้ข้อมูล dual polarimetric radar ข้อมูลจากกรมอุตุนิยมวิทยา ประเทศไทย เพื่อศึกษาโครงสร้างของเมฆฝนที่เกิดขึ้น วิเคราะห์การเคลื่อนที่ของมวลอากาศจากข้อมูล reanalysis ERA5 ทำการสร้างข้อมูลจำลองการหยั่งอากาศจาก ERA5 ในพื้นที่ใกล้เคียงพื้นที่ประสบภัย เพื่อนำมาใช้ในการจำแนกหยาดน้ำฟ้าในช่วงเวลาเกิดเหตุ แล้วทำการวิเคราะห์ลักษณะของภูมิประเทศที่สัมพันธ์กับลมและการเคลื่อนที่ของกลุ่มเมฆฝน การจำแนกหยาดน้ำฟ้าใช้วิธีการของ CSU Tool แล้วใช้ PyTDA และ Single dop ในการวิเคราะห์กระแสลม หากมีข้อสงสัยประการใด ติดต่อ ผู้ช่วยศาสตราจารย์ ดร.นัฐพล มหาวิค มหาวิทยาลัยนเรศวร nattaponm@nu.ac.th